Add support for querying properties and output groups
This change adds support for querying MSBuild properties and output groups. A couple of new menu items, "Get MSBuild Property..." and "Get Output Group...", have been added to the Tools menu. Selecting one of these pops up a dialog box where the name of a property or output group can be entered. The results are then displayed in a dedicated pane in the Output Window.

In the current implementation we query each project in the solution, rather than having the user pick one. Querying for MSBuild properties goes through the `IVsBuildPropertyStorage` interface, and the output groups are retrieved through the `EnvDTE.ConfigurationManager` interface.