Limit package to VS 2017 and 2019
Previously, the supported version range was unbounded. This meant the package would advertise itself as supporting VS2022. Due to the significant changes in VS2022 this version of the package does not work in VS2022. We will create a release branch and have separate packages for VS2017/2019 and VS2022.

This commit is part of that change.