Fix broken search function
At some point in the history, we updated the vs-threading analyzers which flagged some API usage are requiring the main thread. To suppress those, we added `ThreadHelper.ThrowIfNotOnUIThread()`.

In reality, these search methods are called on a worker thread. This led to an exception, which caused search to fail.

The fix is to remove these checks, suppress the warnings, and document the behaviour.