Allow "elapsed" column value to be copied to the clipboard
Among the columns of the build logging table, the elapsed column was unique in how it provided values to the UI. The other columns provided a string value, while the elapsed column returned a `FrameworkElement` in order to right-align the column.

When copying the value to the clipboard, the platform's control doesn't know how to convert the `TextBlock` element into a string, so skips its value in the output.

The solution is to override both the string-providing method and the FrameworkElement-providing method.