Disable table menu items when nothing selected
This change made in response to confusion about the correct use of the UI seen on Twitter:

https://twitter.com/archiecoder/status/1598485876518096896

The user is trying to save a log file, but nothing is selected. We showed menu items that do nothing when clicked.

With this change, the menu items are only displayed when something will actually occur.