Validation won't fail the build
Fixes #467

If we receive a build finish notification more than once, don't throw an exception.

Unfortunately we don't know why this would occur, but it reportedly happened during a publish operation.

In such cases, it's not clear that we can do anything more useful than logging the error via tracing and continuing.